completed
completed the missing

+++ app/Http/Controllers/OrganizationController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Models\Organization;
 use App\Http\Controllers\AuthController;
+use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
+use Illuminate\Support\Arr;
 
 class OrganizationController extends Controller
 {
@@ -16,7 +19,17 @@ class OrganizationController extends Controller
      */
     public function index()
     {
-        $organizations = Organization::all();
+      //  dump($organizations= $users = DB::table('organizations')->where('user_id', '=',  Auth::id())->get());
+        $this->authorize('viewAny', Organization::class);
+        if (Auth::user()->role == 'Admin'){
+            $organizations = Organization::all();
+        }
+        if (Auth::user()->role == 'Employer'){
+            $organizations=DB::table('organizations')->whereUser_idAndDeleted_at(Auth::id(),null)->get();
+        }
+
+
+
         return response()->json($organizations);
     }
 
@@ -39,16 +52,19 @@ class OrganizationController extends Controller
     public function store(Request $request)
     {
 
-        //dd(auth()->check());
-      // $organization = Organization::create($request->all());
+        //$this->authorize('create', Organization::class);
+        if (Auth::user()->role == 'Employer') {
+
         $organization = Organization::make();
-        $organization->user_id = auth()->id(); //Auth::user()->id;
+        $organization->user_id = auth()->id();
         $organization->title = $request->title;
         $organization->country = $request->country;
         $organization->city = $request->city;
         $organization->save();
 
         return response()->json($organization);
+        }
+        return response()->json('Only Employers can store organizations');
 
     }
 
@@ -60,7 +76,50 @@ class OrganizationController extends Controller
      */
     public function show(Organization $organization)
     {
-        //
+
+        $this->authorize('view', $organization);
+        $vacancies = $_GET['vacancies'];
+        $workers = $_GET['workers'];
+        $result=collect ();
+        if ($vacancies == 1) {
+
+           $result= DB::table('vacancies')
+           ->whereDeleted_atAndOrganization_id(null,$organization->id)
+           ->whereColumn('workers_amount', '>' , 'workers_apply_amount')
+           ->get()
+           ->prepend('list of active vacancies');
+        }
+
+        if ($vacancies == 2) {
+
+             $result= DB::table('vacancies')
+           ->whereDeleted_atAndOrganization_id(null,$organization->id)
+           ->whereColumn('workers_amount', '<=' , 'workers_apply_amount')
+           ->get()
+           ->prepend('list of closed (full) vacancies');
+        }
+
+        if ($vacancies == 3) {
+
+
+            $result=DB::table('vacancies')
+            ->whereOrganization_idAndDeleted_at($organization->id,null)->get()->prepend('list of all organization vacancies');
+        }
+
+        if ($workers == 1) {
+            $vacancies=DB::table('vacancies')->whereOrganization_idAndDeleted_at($organization->id,null)->pluck('id');
+            $array = Arr::flatten($vacancies);
+            $item=collect();
+            foreach ($array as $value) {
+                $item->push(DB::table('user_vacancy')->whereVacancy_id($value)->pluck('user_id'));
+            }
+
+            $flattened = Arr::flatten($item);
+            $result->push('Booked workers');
+            $result->push(User::find($flattened));
+
+        }
+        return response()->json($result);
     }
 
     /**
@@ -83,6 +142,7 @@ class OrganizationController extends Controller
      */
     public function update(Request $request, Organization $organization)
     {
+        $this->authorize('update', $organization);
         $organization->update($request->all());
         return response()->json($organization);
     }
@@ -95,7 +155,27 @@ class OrganizationController extends Controller
      */
     public function destroy(Organization $organization)
     {
+        $this->authorize('delete', $organization);
         $organization->delete();
         return response()->json(['message'=> 'organization with id: '.$organization->id.' successfully deleted']);
+    }
+
+
+    /**
+     * statsOrganization.
+     *
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function statsOrganization(){
+        if (Auth::user()->role == 'Admin') {
+            $active= count(DB::table('organizations')->whereDeleted_at(null)->get());
+            $softdeleted= count(DB::table('organizations')->where('deleted_at','!=', null)->get());
+            $all= count(DB::table('organizations')->get());
+
+
+            return response()->json(['Active organizations : ' =>$active, 'SoftDeleted organizations: ' =>$softdeleted, 'All organizations: '=>$all]);
+        }
+        return response()->json('Only Admin authorized for this action');
     }
 }
