+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\User;
 use App\Http\Resources\UserResource;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Exception;
 
 class UserController extends Controller
@@ -18,8 +19,17 @@ class UserController extends Controller
     public function index()
     {
         $this->authorize('viewAny', User::class);
-       $users = User::all();
-       return response()->json($users);
+    $term = $_GET['search'];
+    $result =User::where('first_name', 'like', "%{$term}%")
+            ->orWhere('last_name', 'like', "%{$term}%")
+            ->orWhere('city', 'like', "%{$term}%")
+            ->orWhere('country', 'like', "%{$term}%")
+            ->get();
+
+
+ return response()->json($result);
+      // $users = User::all();
+      // return response()->json($users);
     }
 
     /**
@@ -41,7 +51,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        dd(auth()->user());
+        $this->authorize('view', $user);
        return response()->json($user);
     }
 
@@ -54,6 +64,7 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        $this->authorize('update', $user);
         $user->update($request->all());
         return response()->json($user);
     }
@@ -66,7 +77,26 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
         $user->delete();
         return response()->json(['message'=> 'User with id: '.$user->id.' successfully deleted']);
+    }
+
+    /**
+     * statsUser.
+     *
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function statsUser(){
+        if (Auth::user()->role == 'Admin') {
+            $admins= count(DB::table('users')->whereRole('Admin')->get());
+            $employers= count(DB::table('users')->whereRole('Employer')->get());
+            $workers= count(DB::table('users')->whereRole('Worker')->get());
+
+
+            return response()->json(['Admins : ' =>$admins, 'Employers: ' =>$employers, 'Workers: '=>$workers]);
+        }
+        return response()->json('Only Admin authorized for this action');
     }
 }
